requirements.txt
beautifulsoup4==4.12.3
google-auth==2.39.0
gspread==6.1.2
python-dotenv==1.0.1
requests==2.32.3
